+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
     *Kristján Oddsson*
 
+### Accessibility
+
+* Accept `aria-current="true"` in tabbed components.
+
 ### Changes
 
 * Promote `Tooltip` component to beta.